# Conflicts:
#	frontend/web/src/components/diagnostics/StrategyReadinessPanel.tsx
#	frontend/web/src/components/responsive/DesktopThreePaneShell.test.tsx
#	frontend/web/src/routes/eval-runs-detail.test.tsx

+++ crates/xvision-identity/src/contracts.rs
@@ -142,7 +142,7 @@ sol! {
             uint96  sellerProceeds,
             uint96  protocolProceeds,
             uint256 licenseTokenId,
-            uint8   payerKind,
+            uint16  payerKind,  // v1 placeholder (mirrors purchasePath); see Marketplace.sol
             uint8   purchasePath
         );
     }
